Cirq Overlays: lock google_api_core @ 0.16.0
Cirq 0.8 requires fairly recent version of google_api_core, which isn't
compatible with NixOS 19.09. So we overlay that in our python build here
to prevent conflicts. Seems to be working fine.

We also add a function to prevent overriding if the upstream version is
newer. This is in keeping with the purpose of this to provide newer
versions of packages in-tree for older NixOS versions (i.e. if ARTIQ is
still on 19.09)