pythonPackages.qiskit-ibmq-provider: optional visualizations
Break out the visualization packages from the main code, to make them
an optional extra.
This can speed up import time if heavy packages like matplotlib
are not in the import path.
It looks like matplotlib might still be in the closure, so that will
need to be tracked down.